2 unit tests pass

+++ MassSpringSystem/Simulations/RigidBody.h
@@ -7,7 +7,7 @@ class RigidBody {
 public:
 	//Constructor
 	RigidBody(Vec3 position, Vec3 size, float mass);
-	vector<vector<double>> currentInverseTensor();
+	vector<vector<float>> currentInverseTensor();
 
 	//Attributes
 	Vec3 position;
@@ -19,8 +19,8 @@ public:
 	Vec3 torque;
 
 private:
-	vector<vector<double>> tensor;
-	vector<vector<double>> inertiaTensor();
+	vector<vector<float>> tensor;
+	vector<vector<float>> inertiaTensor();
 };
 
 #endif
